+++ src/domain/QuestionTest.java
@@ -16,6 +16,8 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import utils.Status;
+
 /**
  * @author lmrodrigues
  * @author Scheffel-V
@@ -40,26 +42,26 @@ public class QuestionTest {
         tags = null;
         vinicius = new User("vbscheffel", "123abc", "brsufirefox@gmail.com", "Vinicius");
         sujeito = new User("sujeitousername", "123567", "sujeito.legal@gmail.com", "Sujeito");
-        answer = new Answer(10, sujeito, "Text");
-        answer_2 = new Answer(12, sujeito, "Text");
+        answer = new Answer((long) 10, sujeito, "Text");
+        answer_2 = new Answer((long) 12, sujeito, "Text");
         answerList = new ArrayList<Answer>();
         answerList.add(answer);
         answerList.add(answer_2);
-        comment = new Comment(11, sujeito, "Comment text");
-        comment_2 = new Comment(13, sujeito, "Comment text");
+        comment = new Comment((long) 11, sujeito, "Comment text");
+        comment_2 = new Comment((long) 13, sujeito, "Comment text");
         commentList = new ArrayList<Comment>();
         commentList.add(comment);
         commentList.add(comment_2);
         questionList = new ArrayList<Question>();
         questionList.add(new Question()); // 0 - Question null
-        questionList.add(new Question(1, vinicius, "Text", "Title", tags)); // 1
-                                                                            // -
-                                                                            // Question
-                                                                            // one
-        questionList.add(new Question(2, sujeito, "Text Two", "Title Two", tags)); // 2
-                                                                                   // -
-                                                                                   // Question
-                                                                                   // two
+        questionList.add(new Question((long) 1, vinicius, "Text", "Title", tags)); // 1
+        // -
+        // Question
+        // one
+        questionList.add(new Question((long) 2, sujeito, "Text Two", "Title Two", tags)); // 2
+        // -
+        // Question
+        // two
     }
 
     /**
@@ -72,7 +74,8 @@ public class QuestionTest {
         assertNotNull("Question argumentless constructor gives null object.", questionList.get(0));
         assertTrue("Title field of argumentless Question construct is not null.",
                 questionList.get(0).getTitle() == null);
-        assertTrue("Tags field of argumentless Question construct is not null.", questionList.get(0).getTags() == null);
+        assertTrue("Tags field of argumentless Question construct is not null.",
+                questionList.get(0).getTags() == null);
         assertTrue("Answers field of argumentless Question construct is not null.",
                 questionList.get(0).getAnswers() == null);
         assertTrue("Comments field of argumentless Question construct is not null.",
@@ -90,17 +93,24 @@ public class QuestionTest {
      */
     @Test
     public final void testQuestionWithArguments() {
-        assertNotNull("Question constructor with arguments gives null object.", questionList.get(1));
-        assertTrue("Problem in ID field of construct with arguments.", questionList.get(1).getId() == 1);
-        assertTrue("Problem in author field of construct with arguments.", questionList.get(1).getAuthor() == vinicius);
-        assertTrue("Problem in text field of construct with arguments.", questionList.get(1).getText() == "Text");
-        assertTrue("Problem in title field of construct with arguments.", questionList.get(1).getTitle() == "Title");
-        assertTrue("Problem in tags field of construct with arguments.", questionList.get(1).getTags() == tags);
-        assertEquals("Problem in Answers list field from construct with arguments.", new ArrayList<Answer>(),
-                questionList.get(1).getAnswers());
-        assertEquals("Problem in Comments list field from construct with arguments.", new ArrayList<Comment>(),
-                questionList.get(1).getComments());
-        assertNull("bestAnswer field of construct with arguments is not null.", questionList.get(1).getBestAnswer());
+        assertNotNull("Question constructor with arguments gives null object.",
+                questionList.get(1));
+        assertTrue("Problem in ID field of construct with arguments.",
+                questionList.get(1).getId() == 1);
+        assertTrue("Problem in author field of construct with arguments.",
+                questionList.get(1).getAuthor() == vinicius);
+        assertTrue("Problem in text field of construct with arguments.",
+                questionList.get(1).getText() == "Text");
+        assertTrue("Problem in title field of construct with arguments.",
+                questionList.get(1).getTitle() == "Title");
+        assertTrue("Problem in tags field of construct with arguments.",
+                questionList.get(1).getTags() == tags);
+        assertEquals("Problem in Answers list field from construct with arguments.",
+                new ArrayList<Answer>(), questionList.get(1).getAnswers());
+        assertEquals("Problem in Comments list field from construct with arguments.",
+                new ArrayList<Comment>(), questionList.get(1).getComments());
+        assertNull("bestAnswer field of construct with arguments is not null.",
+                questionList.get(1).getBestAnswer());
         assertSame("Problem in status field from construct with arguments.", Status.OPEN,
                 questionList.get(1).getStatus());
     }
@@ -123,9 +133,9 @@ public class QuestionTest {
     @Test
     public final void testSetId() {
 
-        questionList.get(1).setId(2);
+        questionList.get(1).setId((long) 2);
         assertTrue("Problem in setId of Answer construct.", questionList.get(1).getId() == 2);
-        questionList.get(1).setId(3);
+        questionList.get(1).setId((long) 3);
         assertTrue("Problem in setId of Answer construct.", questionList.get(1).getId() == 3);
     }
 
@@ -137,7 +147,8 @@ public class QuestionTest {
     public final void testGetDate() {
         Date date = new Date();
         questionList.get(1).setDate(date);
-        assertTrue("Problem in getDate of Answer construct.", questionList.get(1).getDate() == date);
+        assertTrue("Problem in getDate of Answer construct.",
+                questionList.get(1).getDate() == date);
     }
 
     /**
@@ -149,9 +160,11 @@ public class QuestionTest {
         Date date_1 = new Date();
         Date date_2 = new Date();
         questionList.get(1).setDate(date_1);
-        assertTrue("Problem in setDate of Answer construct.", questionList.get(1).getDate() == date_1);
+        assertTrue("Problem in setDate of Answer construct.",
+                questionList.get(1).getDate() == date_1);
         questionList.get(1).setDate(date_2);
-        assertTrue("Problem in setDate of Answer construct.", questionList.get(1).getDate() == date_2);
+        assertTrue("Problem in setDate of Answer construct.",
+                questionList.get(1).getDate() == date_2);
     }
 
     /**
@@ -162,10 +175,12 @@ public class QuestionTest {
     public final void testSetAuthor() {
 
         questionList.get(1).setAuthor(vinicius);
-        assertTrue("Problem in setAuthor of Answer construct.", questionList.get(1).getAuthor() == vinicius);
+        assertTrue("Problem in setAuthor of Answer construct.",
+                questionList.get(1).getAuthor() == vinicius);
         User user_1 = new User();
         questionList.get(1).setAuthor(user_1);
-        assertTrue("Problem in setAuthor of Answer construct.", questionList.get(1).getAuthor() == user_1);
+        assertTrue("Problem in setAuthor of Answer construct.",
+                questionList.get(1).getAuthor() == user_1);
     }
 
     /**
@@ -175,10 +190,12 @@ public class QuestionTest {
     @Test
     public final void testGetAuthor() {
 
-        assertTrue("Problem in getAuthor of Answer construct.", questionList.get(1).getAuthor() == vinicius);
+        assertTrue("Problem in getAuthor of Answer construct.",
+                questionList.get(1).getAuthor() == vinicius);
         User user_1 = new User();
         questionList.get(1).setAuthor(user_1);
-        assertTrue("Problem in getAuthor of Answer construct.", questionList.get(1).getAuthor() == user_1);
+        assertTrue("Problem in getAuthor of Answer construct.",
+                questionList.get(1).getAuthor() == user_1);
     }
 
     /**
@@ -188,9 +205,11 @@ public class QuestionTest {
     @Test
     public final void testGetText() {
 
-        assertTrue("Problem in getText of Answer construct.", questionList.get(1).getText() == "Text");
+        assertTrue("Problem in getText of Answer construct.",
+                questionList.get(1).getText() == "Text");
         questionList.get(1).setText("Text Test");
-        assertTrue("Problem in getText of Answer construct.", questionList.get(1).getText() == "Text Test");
+        assertTrue("Problem in getText of Answer construct.",
+                questionList.get(1).getText() == "Text Test");
     }
 
     /**
@@ -201,9 +220,11 @@ public class QuestionTest {
     public final void testSetText() {
 
         questionList.get(1).setText("Text");
-        assertTrue("Problem in setText of Answer construct.", questionList.get(1).getText() == "Text");
+        assertTrue("Problem in setText of Answer construct.",
+                questionList.get(1).getText() == "Text");
         questionList.get(1).setText("Text Test");
-        assertTrue("Problem in setText of Answer construct.", questionList.get(1).getText() == "Text Test");
+        assertTrue("Problem in setText of Answer construct.",
+                questionList.get(1).getText() == "Text Test");
     }
 
     /**
@@ -255,7 +276,8 @@ public class QuestionTest {
     @Test
     public final void testGetAnswers() {
         assertTrue("Problem in getAnswers.", questionList.get(0).getAnswers() == null);
-        assertEquals("Problem in getAnswers.", questionList.get(1).getAnswers(), new ArrayList<Answer>());
+        assertEquals("Problem in getAnswers.", questionList.get(1).getAnswers(),
+                new ArrayList<Answer>());
         questionList.get(1).setAnswers(answerList);
         assertTrue("Problem in getAnswers.", questionList.get(1).getAnswers().get(0).getId() == 10);
         assertTrue("Problem in getAnswers.", questionList.get(1).getAnswers().get(1).getId() == 12);
@@ -312,10 +334,13 @@ public class QuestionTest {
     @Test
     public final void testGetComments() {
         assertTrue("Problem in getComments.", questionList.get(0).getComments() == null);
-        assertEquals("Problem in getComments.", questionList.get(1).getComments(), new ArrayList<Comment>());
+        assertEquals("Problem in getComments.", questionList.get(1).getComments(),
+                new ArrayList<Comment>());
         questionList.get(1).setComments(commentList);
-        assertTrue("Problem in getComments.", questionList.get(1).getComments().get(0).getId() == 11);
-        assertTrue("Problem in getComments.", questionList.get(1).getComments().get(1).getId() == 13);
+        assertTrue("Problem in getComments.",
+                questionList.get(1).getComments().get(0).getId() == 11);
+        assertTrue("Problem in getComments.",
+                questionList.get(1).getComments().get(1).getId() == 13);
     }
 
     /**
@@ -325,8 +350,10 @@ public class QuestionTest {
     @Test
     public final void testSetComments() {
         questionList.get(1).setComments(commentList);
-        assertTrue("Problem in setComments.", questionList.get(1).getComments().get(0).getId() == 11);
-        assertTrue("Problem in setComments.", questionList.get(1).getComments().get(1).getId() == 13);
+        assertTrue("Problem in setComments.",
+                questionList.get(1).getComments().get(0).getId() == 11);
+        assertTrue("Problem in setComments.",
+                questionList.get(1).getComments().get(1).getId() == 13);
         questionList.get(1).setComments(null);
         assertTrue("Problem in setComments.", questionList.get(1).getComments() == null);
     }
@@ -338,10 +365,13 @@ public class QuestionTest {
     @Test
     public final void testAddComment() {
         questionList.get(2).addComment(comment_2);
-        assertTrue("Problem in addComment.", questionList.get(2).getComments().get(0).getId() == 13);
+        assertTrue("Problem in addComment.",
+                questionList.get(2).getComments().get(0).getId() == 13);
         questionList.get(2).addComment(comment);
-        assertTrue("Problem in addComment.", questionList.get(2).getComments().get(0).getId() == 13);
-        assertTrue("Problem in addComment.", questionList.get(2).getComments().get(1).getId() == 11);
+        assertTrue("Problem in addComment.",
+                questionList.get(2).getComments().get(0).getId() == 13);
+        assertTrue("Problem in addComment.",
+                questionList.get(2).getComments().get(1).getId() == 11);
     }
 
     /**
@@ -352,14 +382,16 @@ public class QuestionTest {
     public final void testDelComment() {
         questionList.get(1).setComments(commentList);
         questionList.get(1).delComment(comment_2);
-        assertTrue("Problem in delComment.", questionList.get(1).getComments().get(0).getId() == 11);
+        assertTrue("Problem in delComment.",
+                questionList.get(1).getComments().get(0).getId() == 11);
         questionList.get(1).delComment(comment);
         assertTrue("Problem in delComment.", questionList.get(1).getComments().isEmpty() == true);
         questionList.get(1).delComment(comment);
         assertTrue("Problem in delComment.", questionList.get(1).getComments().isEmpty() == true);
         questionList.get(1).addComment(comment);
         questionList.get(1).delComment(comment_2);
-        assertTrue("Problem in delComment.", questionList.get(1).getComments().get(0).getId() == 11);
+        assertTrue("Problem in delComment.",
+                questionList.get(1).getComments().get(0).getId() == 11);
     }
 
     /**
@@ -395,11 +427,12 @@ public class QuestionTest {
         assertTrue("Problem in getStatus method.", questionList.get(0).getStatus() == null);
         assertTrue("Problem in getStatus method.", questionList.get(1).getStatus() == Status.OPEN);
         questionList.get(1).setStatus(Status.CLOSED);
-        assertTrue("Problem in getStatus method.", questionList.get(1).getStatus() == Status.CLOSED);
+        assertTrue("Problem in getStatus method.",
+                questionList.get(1).getStatus() == Status.CLOSED);
     }
 
     /**
-     * Test method for {@link domain.Question#setStatus(domain.Status)}. This
+     * Test method for {@link domain.Question#setStatus(utils.Status)}. This
      * test will check the setStatus method.
      */
     @Test
@@ -407,7 +440,8 @@ public class QuestionTest {
         questionList.get(1).setStatus(Status.OPEN);
         assertTrue("Problem in setStatus method.", questionList.get(1).getStatus() == Status.OPEN);
         questionList.get(1).setStatus(Status.CLOSED);
-        assertTrue("Problem in setStatus method.", questionList.get(1).getStatus() == Status.CLOSED);
+        assertTrue("Problem in setStatus method.",
+                questionList.get(1).getStatus() == Status.CLOSED);
     }
 
     /**
@@ -417,12 +451,15 @@ public class QuestionTest {
     @Test
     public final void testOpenQuestion() {
         questionList.get(1).openQuestion();
-        assertTrue("Problem in openQuestion method.", questionList.get(1).getStatus() == Status.OPEN);
+        assertTrue("Problem in openQuestion method.",
+                questionList.get(1).getStatus() == Status.OPEN);
         questionList.get(1).openQuestion();
-        assertTrue("Problem in openQuestion method.", questionList.get(1).getStatus() == Status.OPEN);
+        assertTrue("Problem in openQuestion method.",
+                questionList.get(1).getStatus() == Status.OPEN);
         questionList.get(1).closeQuestion();
         questionList.get(1).openQuestion();
-        assertTrue("Problem in openQuestion method.", questionList.get(1).getStatus() == Status.OPEN);
+        assertTrue("Problem in openQuestion method.",
+                questionList.get(1).getStatus() == Status.OPEN);
     }
 
     /**
@@ -432,12 +469,15 @@ public class QuestionTest {
     @Test
     public final void testCloseQuestion() {
         questionList.get(1).closeQuestion();
-        assertTrue("Problem in closeQuestion method.", questionList.get(1).getStatus() == Status.CLOSED);
+        assertTrue("Problem in closeQuestion method.",
+                questionList.get(1).getStatus() == Status.CLOSED);
         questionList.get(1).closeQuestion();
-        assertTrue("Problem in closeQuestion method.", questionList.get(1).getStatus() == Status.CLOSED);
+        assertTrue("Problem in closeQuestion method.",
+                questionList.get(1).getStatus() == Status.CLOSED);
         questionList.get(1).openQuestion();
         questionList.get(1).closeQuestion();
-        assertTrue("Problem in closeQuestion method.", questionList.get(1).getStatus() == Status.CLOSED);
+        assertTrue("Problem in closeQuestion method.",
+                questionList.get(1).getStatus() == Status.CLOSED);
     }
 
 }
